backend/requirements.txt
Pillow==10.4.0
PyJWT==2.8.0
PyMuPDF==1.24.9
SQLAlchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
boto3==1.34.162
celery==5.3.6
fastapi==0.111.0
jinja2==3.1.4
markdown==3.6
minio==7.2.7
passlib==1.7.4
bcrypt==4.0.1
pgvector==0.2.4
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
pydantic==2.8.2
pytesseract==0.3.10
python-docx==0.8.11
python-multipart==0.0.9
pyyaml==6.0.2
redis==5.0.7
reportlab==4.2.2
requests==2.32.3
stripe>=9.0.0
tenacity==8.3.0
uvicorn[standard]==0.30.1
openai>=1.0.0
anthropic>=0.34.0